+++ utils/buildTests.js
@@ -13,3 +13,7 @@ fs.writeFileSync('./site/tests/mhchem-tests.html', mhchemTests)
 let mozillaTests = fs.readFileSync('./test/mozilla-tests.md').toString('utf8')
 mozillaTests = hurmetMark.hmd.md2html(mozillaTests, true)
 fs.writeFileSync('./site/tests/mozilla-tests.html', mozillaTests)
+
+let wikiTests = fs.readFileSync('./test/wiki-tests.md').toString('utf8')
+wikiTests = hurmetMark.hmd.md2html(wikiTests, true)
+fs.writeFileSync('./site/tests/wiki-tests.html', wikiTests)